requirements/base.txt
Django==1.11.11
django-allauth>=0.35.0
django-braces>=1.12.0
django-cors-headers>=2.1.0
django-environ>=0.4.4
django-extensions>=2.0.0
django-markupfield==1.4.3
django-model-utils>=3.1.1
django-silk>=2.0.0
django-solo>=1.1.3

djangorestframework>=3.7.7
djangorestframework-jwt>=1.11.0
django-rest-auth>=0.9.3

coreapi>=2.3.3
docutils==0.14
markdown==2.6.11
Pillow>=5.0.0
psycopg2==2.7.3.2
Pygments>=2.2.0
pytz>=2018.3
raven>=6.5.0
requests>=2.18.4
sqlformatter>=1.3
sqlparse>=0.2.4
